+++ CHANGELOG.md
@@ -6,7 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-[Unreleased]: https://github.com/trussed-dev/admin-app/compare/0.3.0...HEAD
+[Unreleased]: https://github.com/trussed-dev/admin-app/compare/0.4.0...HEAD
+
+-
+
+## [0.4.0] 2026-09-16
+
+[0.4.0]: https://github.com/trussed-dev/admin-app/compare/0.3.0...0.4.0
 
 - Make `Reboot::reboot_to_firmware_update` non-diverging.
 - Add `Data` struct and change `App::load_config` and `App::with_default_config` to use it.